+++ packages/main/src/TabContainer.hbs
@@ -36,12 +36,7 @@
 						</li>
 					{{/unless}}
 					{{#if this.isSeparator}}
-						<li id="{{this._id}}"
-							tabindex="{{this._tabIndex}}"
-							role="separator"
-							class="{{../classes.separator}}"
-						>
-						</li>
+						<li id="{{this._id}}" role="separator" class="{{../classes.separator}}"></li>
 					{{/if}}
 				{{/each}}
 			</ul>
