+++ packages/main/src/TabSeparator.hbs
@@ -1,1 +1,1 @@
-<div {{> controlData}} role="separator" class="{{classes.main}}"></div>
+<li {{> controlData}} role="separator"></li>
